+++ src/genre/list.rs
@@ -1,5 +1,7 @@
 use std::borrow::Cow;
 
+use crate::client::Executor;
+
 use super::Genre;
 
 const TV_PATH: &str = "/genre/tv/list";
@@ -50,7 +52,10 @@ impl crate::prelude::Command for GenreList {
         }
     }
 
-    async fn execute(&self, client: &crate::Client) -> Result<Self::Output, crate::error::Error> {
+    async fn execute<E: Executor>(
+        &self,
+        client: &crate::Client<E>,
+    ) -> Result<Self::Output, crate::error::Error> {
         client
             .execute::<GenreResult>(self.path().as_ref(), self.params())
             .await
@@ -62,15 +67,16 @@ impl crate::prelude::Command for GenreList {
 mod tests {
     use mockito::Matcher;
 
+    use crate::client::reqwest::ReqwestExecutor;
+    use crate::client::Client;
     use crate::prelude::Command;
-    use crate::Client;
 
     use super::GenreList;
 
     #[tokio::test]
     async fn movie_works() {
         let mut server = mockito::Server::new_async().await;
-        let client = Client::builder()
+        let client = Client::<ReqwestExecutor>::builder()
             .with_api_key("secret".into())
             .with_base_url(server.url())
             .build()
@@ -92,7 +98,7 @@ mod tests {
     #[tokio::test]
     async fn tv_works() {
         let mut server = mockito::Server::new_async().await;
-        let client = Client::builder()
+        let client = Client::<ReqwestExecutor>::builder()
             .with_api_key("secret".into())
             .with_base_url(server.url())
             .build()
@@ -114,7 +120,7 @@ mod tests {
     #[tokio::test]
     async fn invalid_api_key() {
         let mut server = mockito::Server::new_async().await;
-        let client = Client::builder()
+        let client = Client::<ReqwestExecutor>::builder()
             .with_api_key("secret".into())
             .with_base_url(server.url())
             .build()
@@ -131,13 +137,13 @@ mod tests {
             .await;
         let err = cmd.execute(&client).await.unwrap_err();
         let server_err = err.as_server_error().unwrap();
-        assert_eq!(server_err.body.as_other_error().unwrap().status_code, 7);
+        assert_eq!(server_err.status_code, 7);
     }
 
     #[tokio::test]
     async fn resource_not_found() {
         let mut server = mockito::Server::new_async().await;
-        let client = Client::builder()
+        let client = Client::<ReqwestExecutor>::builder()
             .with_api_key("secret".into())
             .with_base_url(server.url())
             .build()
@@ -154,21 +160,22 @@ mod tests {
             .await;
         let err = cmd.execute(&client).await.unwrap_err();
         let server_err = err.as_server_error().unwrap();
-        assert_eq!(server_err.body.as_other_error().unwrap().status_code, 34);
+        assert_eq!(server_err.status_code, 34);
     }
 }
 
 #[cfg(all(test, feature = "integration"))]
 mod integration_tests {
+    use crate::client::reqwest::ReqwestExecutor;
+    use crate::client::Client;
     use crate::prelude::Command;
-    use crate::Client;
 
     use super::GenreList;
 
     #[tokio::test]
     async fn execute_tv() {
         let secret = std::env::var("TMDB_TOKEN_V3").unwrap();
-        let client = Client::new(secret);
+        let client = Client::<ReqwestExecutor>::new(secret);
         let mut cmd = GenreList::tv();
         cmd.language = Some("en-US".into());
 
@@ -179,7 +186,7 @@ mod integration_tests {
     #[tokio::test]
     async fn execute_movie() {
         let secret = std::env::var("TMDB_TOKEN_V3").unwrap();
-        let client = Client::new(secret);
+        let client = Client::<ReqwestExecutor>::new(secret);
         let mut cmd = GenreList::movie();
         cmd.language = Some("en-US".into());